tegra234-flash-dry: Fix generating boot0.img in Jenkins
The boot0.img was generated incorrectly because the
signed binaries were not present at that point. This caused
failure in booting after the update. The issue was not visible
on a local yocto setup which had multiple builds done because
the files were already deployed from previous tegra234-flash-dry
runs.

Changelog-entry: tegra234-flash-dry: Fix generating boot0.img in Jenkins
Signed-off-by: Alexandru Costache <alexandru@balena.io>