edk2-firmware-tegra: Load dtb from root partition if unspecified in env
Previously, the uefi dtb was used if none was specified in extra_uEnv.txt.
However, when rolling back to L4T 35.1.0 we disable xudc from the
device-tree upon hup, to avoid a panic in the old rootfs, and thus
need to load the rootfs dtb instead of the one provided by the firmware.

Signed-off-by: Alexandru Costache <alexandru@balena.io>