hostapp-update-hooks: Update dependencies - see NOTE
The Xavier NX eMMC custom device example
uses a different recipe for signed tegra binaries,
other than the Orin devices, therefore we need to
add the right dependency/

IMPORTANT NOTE: This branch should be used as an example
for development and for testing a custom carrier board.
It is incomplete, untested and should not be used for
production devices.

Signed-off-by: Alexandru Costache <alexandru@balena.io>