hostapp-update-hooks: Move efivars trigger to 'after' hook
In PR https://github.com/balena-os/balena-jetson-orin/pull/238
we added the prebuilt capsule binaries as well as the legacy
bundle for L4T 35.3.1, and we also added the switch of the esp
and resin-boot partition types - which is performed if resin-boot
does not have the esp,boot flags set by the flashing tools. These
flags are needed by the UEFI firmware to load the capsule from
the resin-boot partition, and we'll modify jetson-flash to set
them during provisioning from now on.

The uefi variable setting approach in the above PR won't work
as-is because balenaHUP does not run hostapp-update in a container,
and because the existing hostapp-update script in the current releases
does not bindmount the sysfs directory. Thus it doesn't allow for the
efivars needed to trigger the capsule update to be written.

The approach we want to take further and implement here is:

  - use legacy update if updating from a very old release which
    does not support capsule update
  - trigger the capsule update from the after hook - which runs in
    the current OS - until a release with an update hostapp-update
    script becomes available in production for the AGX orin. See
    meta-balena PR #3194
  - once a release that bindmounts sysfs is available, we'll switch
    to using the legacy mode for any releases that don't have the new
    hostapp-update script, and remove the capsule update from the after hook

The advantage of the above approach is that we create an update path
that is a bit faster and safer than the legacy update mode, until a
newer hostapp-update script is available in production images.

Changelog-entry: hostapp-update-hooks: Move efivars trigger to 'after' hook
Signed-off-by: Alexandru Costache <alexandru@balena.io>