linux-tegra: On the Orin AGX use rtc0 as system clock
The Orin AGX devkit has two RTC clocks, one in the SOM and one in the
carrier board. The BSP defaults to using the one in the SOM but only
the one in the carrier board is connected to a backup battery and is able
to keep time between reboots.

Changelog-entry: Use battery backed-up RTC in Orin AGX devkit
Signed-off-by: Alex Gonzalez <alexg@balena.io>