linux-tegra: Rename AGX Orin Devkit 32GB SPI enable dtb
Adds a custom DTB that enables the SPI pins in the 40-pin header of the
Jetson AGX Orin Dev Kit. The 40-pin header defaults to GPIO. This DTB
was generated using the jetson-io utility that is supplied with L4T and
JetPack to configure the 40-pin header. The DTB has been verified by
uploading to balenaOS devices and using the device configuration in
balenaCloud.

Changelog-entry: Add SPI support for 40-pin header
Signed-off-by: Christopher Field <chris.field@theiascientific.com>