hostapp-update-hooks: Move uefi variable writing to the current OS hook
... for the AGX Orin and ensure there's no leftover capsule
in the boot partition, which could potentially cause rollbacks
to fail due to lack of space. In this commit we also extract the
UEFI capsule to the esp partition, because with recent BSP updates
it has become larger than what can fit into the boot partition.

Signed-off-by: Alexandru Costache <alexandru@balena.io>
Changelog-entry: hostapp-update-hooks: Move uefi variable writing to the current OS hook