hostapp-update-hooks: Fallback to legacy QSPI update mode
... if alternate slot is unbootable on AGX Orin

During rollbacks tests the alternate slot might become unbootable
when applying capsule updates from 35.4.1 to 35.5.0 and vice-versa,
and in this case future capsule updates will apply but fail to boot.

In this case the only solution is to fallback to re-writing
the entire QSPI during boot, with the new 35.5.0 boot firmware.

Changelog-entry: hostapp-update-hooks: Fallback to legacy QSPI update mode if alternate slot is unbootable
Signed-off-by: Alexandru Costache <alexandru@balena.io>