jetson-dtbs: Don't try to install unavailable dtb
Starting with L4T 35.5.0 the jetson-agx-orin-devkit yocto
machine from meta-tegra only specifies and builds the Developer Kit dtb,
without the production module. That module now has a separate
yocto machine. Since we support and test the Devkit with SKU0,
and not hte 32GB produciton module with SKU4, we will continue
to ship the default dtb for this configuration.

Signed-off-by: Alexandru Costache <alexandru@balena.io>