hostapp-update-hooks: Allow reprovisioning the QSPI if alternate slot becomes unbootable
During rollbacks we trigger the UEFI capsule update from l4T 35.5.0
back to to L4T 35.4.1. The UEFI firmware does not allow this by default,
so we patched it to accept downgrades when checking versions.

However, the alternate slot may become unbootable in this process,
as reported by the MB2 precompiled bootloader
(see AGX Orin comments in HUP hook). In this case, we re-write the
entire QSPI on the next HUP, and start off with a
clean-flashed QSPI at the latest version 35.5.0.

Apart from this we unload spi modules, to allow for future
rollbacks from L4T 36.3 to be able to write the necessary
UEFI variables necessary for triggering the rollback
UEFI capsule update.

Changelog-entry: hostapp-update-hooks: Allow reprovisioning the QSPI if alternate slot becomes unbootable
Signed-off-by: Alexandru Costache <alexandru@balena.io>