recipes-bsp/uefi: Update integration for L4T 36.3
... and also switch from building edk2-firmware-tegra
in Yocto to building it in a container. This because
the upstream integration is made to work with Poky
Scarthgap, which is not yet available in meta-balena.

We also updated our balenaOS integration patches for
all boards to allow most rollbacks to work when migrating
from Jetpack 5 to Jetpack 6.

The tested rollback scenarios are in the following order, without re-flashing between steps:

- Perform HUP from Jetpack 5 L4T 35.5.0 and interrupt UEFI capsule update in firmware
- Perform HUP from Jetpack 5 L4T 35.5.0 and trigger rollback altboot
- Perform HUP from Jetpack 5 L4T 35.5.0 and trigger rollback health, using the same device
- Perform HUP to L4T 36.3
- Perform one more HUP to L4T 36.3

Signed-off-by: Alexandru Costache <alexandru@balena.io>