jetson-dtbs: Update default device-tree names in Jetpack 6, drop AGX Orin SPI dtb
The SPI device-tree is deleted for now because it cannot be used
for booting Jetpack 6, since Jetpack 5 device-trees are incompatible
with this new release.

Thus, the dtb will need to be re-created and added back for the new
L4T 36.3.

Changelog-entry: jetson-dtbs: Update default device-tree names in Jetpack 6, drop AGX Orin SPI dtb
Signed-off-by: Alexandru Costache <alexandru@balena.io>