recipes-bsp: Update AGX Orin 32GB UEFI capsule and firmware
This includes an updated tos-optee_t234 and tos.img
which erase the var store when the integrity check fails,
allowing for the system to recover in this case and avoid
bricking. The erasing is performed after the capsule update
is applied, to prevent the system from booting with
L4T 35.5.0 firmware and 36.3 bootloader, kernel and os.

Signed-off-by: Alexandru Costache <alexandru@balena.io>