.gitmodules: update poky branch
We are maintaining a fork of Poky and using a branch there as source
as we need to backport a bitbake patch that is still not contained
in the current Yocto LTS release.

The patch in question is:
https://github.com/balena-os/poky/commit/debd42be214be6bce50a0f0a2fe186b20e6fe6bc

This commit updates the branch to rebase that commit to the current
kirkstone HEAD which includes security and bug fixes.

The old branch is maintained not to loose history, and a new branch has
been preferred over a merge into the current one as that would only
bury the backported patch making it more difficult to follow what the
reasoning for the branch is.

Changelog-entry: Update Poky to kirkstone HEAD
Signed-off-by: Alex Gonzalez <alexg@balena.io>