recipes-bsp/uefi: Force use NewDeviceHierarchy
... even if the device was flashed with NewDeviceHierarchy { data = [00]..
so that it can boot a flasher image when a USB key is inserted.

Not setting this would prevent the UEFI bootloader from the
USB key to be loaded by the UEFI firmware. This is done for the
AGX 64GB only, since it's the only flasher DT which has NewDeviceHierarchy
disabled.

NOTE: Setting the flasher rootfs label from extlinux wouldn't
work as an alternative, because the extlinux.conf file
would be loaded from the eMMC as well.

Signed-off-by: Alexandru Costache <alexandru@balena.io>