recipes-support/jetson-qspi-manager: Switch to super configuration by default
We switch to using the super configuration by default for the Orin Nano Devkit, NX
in Xavier NX Devkit and Seeed J3010, because these carrier boards
do have support for it.

This in turn allows for the UEFI capsule update to apply
the super BPFDDTB in place of the standard one.

We also now permit applying the capsule update during
flashing for boards which have already been provisioned
with Ubuntu JP6.2

Signed-off-by: Alexandru Costache <alexandru@balena.io>