recipes-support/hostapp-update-hooks: Rework hooks for L4T 36.4.3
In this commit we adapt the old 36.3.0 hooks from the inactive rootfs
so that they don't fail during rollback-health.

We also drop the bootslot usage, since that was used to ensure
rollbacks could be performed between JP5 and JP6, and it is no longer
necessary when switching between JP6.0 and JP6.2.

Signed-off-by: Alexandru Costache <alexandru@balena.io>