jetson-dtbs: depend on virtual/dtb
nvidia-kernel-oot-dtb provides virtual/dtb but it's not meant to be
the only provider and even for Balena use cases there might be a need
to use something else, for instance something board-specific.

This recipes works just fine with virtual/dtb:do_deploy dependency
as long as whatever provides virtual/dtb implements do_deploy and
leaves the DTs in the right location.

Signed-off-by: Alex Suykov <alex.suykov@sirinsoftware.com>