Discontinue Forecr DSB ORNX-LAN Orin NX 16GB
This device is not autokitted and will be discontinued
for now. We also remove the Orin Nano 4GB coffee file and
workflow since it never had any release, and it will not be
placed in an autokit nor have any release for the time being.

Changelog-entry: Discontinue Forecr DSB ORNX-LAN Orin NX 16GB
Signed-off-by: Alexandru Costache <alexandru@balena.io>