recipes-bsp/tegra-binaries: Update Seeed J3010 boot blob, drop prebuilt UEFI capsules
The J3010 boot blob has been obtained from a J3010 device
flashed with the same configuration and setup used by
the uefi-capsule-container package.

The flashing line used with the Nvidia setup and
the Seeed overlay is the standard one for NVME provisioning:

sudo ./tools/kernel_flash/l4t_initrd_flash.sh
   --external-device nvme0n1p1 -c tools/kernel_flash/flash_l4t_external.xml
   -p "-c bootloader/generic/cfg/flash_t234_qspi.xml
   --no-systemimg" --network usb0
    jetson-orin-nano-devkit external

and the contents of the jetson-orin-nano-devkit.conf are
the same as the one used by the uefi-capsule-container package.

Apart from this we drop all prebuilt UEFI capsules,
as they are now provided by uefi-capsule-container.

Signed-off-by: Alexandru Costache <alexandru@balena.io>