tegra-binaries/uefi-capsule-container: Create package for building the UEFI capsule
... in a container. Previously building the UEFI capsule worked
only if the entire rootfs was extracted over the BSP archive.
This no longer appears to be necessary, so let's build the UEFI
capsule at runtime.

This capsule contains the uefi firmware patched and built for
balenaOS and is applied during OTA updates, as well during
provisioning with the USB key alone.

Changelog-entry: tegra-binaries/uefi-capsule-container: Create package for building the UEFI capsule
Signed-off-by: Alexandru Costache <alexandru@balena.io>