tegra-binaries/uefi-capsule-container: Create package for building UEFI capsules
... in a container. Previously building the UEFI capsule worked
only if the entire Ubuntu rootfs was extracted over the BSP archive.
This no longer appears to be necessary, so let's build the UEFI
capsule at runtime.

This capsule contains the uefi firmware patched and built for
balenaOS and is applied during OTA updates, as well during
provisioning with the USB key alone.

The jetson_board_spec*.cfg files come from the standard
L4T BSP 36.4.3. In them we comment-out the bup specs which
are not necessary for the corresponding device-type,
in order to keep the resulting UEFI capsule size at
a minimum.

Changelog-entry: tegra-binaries/uefi-capsule-container: Create package for building UEFI capsules
Signed-off-by: Alexandru Costache <alexandru@balena.io>